Create emailsender.py
Please note that it's important to use a proper structure when sending emails to ensure that they are delivered correctly and look professional.

It's also important to use a proper email address and signature for professional communication.

Also, as a best practice it's recommended to use a service like SendGrid, Mailgun, etc. to send mass emails as these services are specifically built for this use case and handle the sending and delivery of the emails.